+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/text/EditInlineMessageText.kt
@@ -2,7 +2,7 @@ package dev.inmo.tgbotapi.extensions.api.edit.text
 
 import dev.inmo.tgbotapi.bot.TelegramBot
 import dev.inmo.tgbotapi.requests.edit.text.EditInlineMessageText
-import dev.inmo.tgbotapi.types.InlineMessageIdentifier
+import dev.inmo.tgbotapi.types.InlineMessageId
 import dev.inmo.tgbotapi.types.LinkPreviewOptions
 import dev.inmo.tgbotapi.types.message.textsources.TextSourcesList
 import dev.inmo.tgbotapi.types.message.ParseMode
@@ -16,7 +16,7 @@ import dev.inmo.tgbotapi.utils.buildEntities
  * as a builder for that
  */
 suspend fun TelegramBot.editMessageText(
-    inlineMessageId: InlineMessageIdentifier,
+    inlineMessageId: InlineMessageId,
     text: String,
     parseMode: ParseMode? = null,
     linkPreviewOptions: LinkPreviewOptions? = null,
@@ -28,7 +28,7 @@ suspend fun TelegramBot.editMessageText(
  * as a builder for that
  */
 suspend fun TelegramBot.editMessageText(
-    inlineMessageId: InlineMessageIdentifier,
+    inlineMessageId: InlineMessageId,
     entities: TextSourcesList,
     linkPreviewOptions: LinkPreviewOptions? = null,
     replyMarkup: InlineKeyboardMarkup? = null
@@ -39,7 +39,7 @@ suspend fun TelegramBot.editMessageText(
  * as a builder for that
  */
 suspend fun TelegramBot.editMessageText(
-    inlineMessageId: InlineMessageIdentifier,
+    inlineMessageId: InlineMessageId,
     separator: TextSource? = null,
     linkPreviewOptions: LinkPreviewOptions? = null,
     replyMarkup: InlineKeyboardMarkup? = null,
@@ -51,7 +51,7 @@ suspend fun TelegramBot.editMessageText(
  * as a builder for that
  */
 suspend fun TelegramBot.editMessageText(
-    inlineMessageId: InlineMessageIdentifier,
+    inlineMessageId: InlineMessageId,
     separator: String,
     linkPreviewOptions: LinkPreviewOptions? = null,
     replyMarkup: InlineKeyboardMarkup? = null,