+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/abstracts/types/InlineMessageAction.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.abstracts.types
 
-import dev.inmo.tgbotapi.types.InlineMessageIdentifier
+import dev.inmo.tgbotapi.types.InlineMessageId
 
 interface InlineMessageAction {
-    val inlineMessageId: InlineMessageIdentifier
+    val inlineMessageId: InlineMessageId
 }