+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InlineMessageIdentifier.kt
@@ -1,0 +1,12 @@
+package dev.inmo.tgbotapi.types
+
+import kotlinx.serialization.Serializable
+import kotlin.jvm.JvmInline
+
+@Serializable
+@JvmInline
+value class InlineMessageId(
+    val string: String
+)
+@Deprecated("Renamed", ReplaceWith("InlineMessageId", "dev.inmo.tgbotapi.types.InlineMessageId"))
+typealias InlineMessageIdentifier = InlineMessageId