+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InlineQueries/ChosenInlineResult/BaseChosenInlineResult.kt
@@ -10,6 +10,6 @@ data class BaseChosenInlineResult(
     override val resultId: InlineQueryIdentifier,
     @SerialName(fromField)
     override val from: User,
-    override val inlineMessageId: InlineMessageIdentifier?,
+    override val inlineMessageId: InlineMessageId?,
     override val query: String
 ) : ChosenInlineResult