+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InlineQueries/ChosenInlineResult/ChosenInlineResult.kt
@@ -1,11 +1,11 @@
 package dev.inmo.tgbotapi.types.InlineQueries.ChosenInlineResult
 
 import dev.inmo.tgbotapi.abstracts.FromUser
-import dev.inmo.tgbotapi.types.InlineMessageIdentifier
+import dev.inmo.tgbotapi.types.InlineMessageId
 import dev.inmo.tgbotapi.types.InlineQueryIdentifier
 
 sealed interface ChosenInlineResult : FromUser {
     val resultId: InlineQueryIdentifier //chosen temporary, can be changed
-    val inlineMessageId: InlineMessageIdentifier?
+    val inlineMessageId: InlineMessageId?
     val query: String
 }