+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/queries/callback/InlineMessageIdCallbackQuery.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.types.queries.callback
 
-import dev.inmo.tgbotapi.types.InlineMessageIdentifier
+import dev.inmo.tgbotapi.types.InlineMessageId
 
 sealed interface InlineMessageIdCallbackQuery : CallbackQuery {
-    val inlineMessageId: InlineMessageIdentifier
+    val inlineMessageId: InlineMessageId
 }