+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/queries/callback/InlineMessageIdDataCallbackQuery.kt
@@ -2,7 +2,6 @@ package dev.inmo.tgbotapi.types.queries.callback
 
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.chat.CommonUser
-import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,6 +9,6 @@ data class InlineMessageIdDataCallbackQuery(
     override val id: CallbackQueryIdentifier,
     override val from: CommonUser,
     override val chatInstance: String,
-    override val inlineMessageId: InlineMessageIdentifier,
+    override val inlineMessageId: InlineMessageId,
     override val data: String
 ) : DataCallbackQuery, InlineMessageIdCallbackQuery